+++ src/model/gp_noise.jl
@@ -1,4 +1,5 @@
-export PowerlawRedNoiseGP, PowerlawDispersionNoiseGP, PowerlawChromaticNoiseGP
+export PowerlawRedNoiseGP,
+    PowerlawDispersionNoiseGP, PowerlawChromaticNoiseGP, calc_noise_weights_inv
 
 function powerlaw(A, γ, f, f1)
     fyr = frequency(1 / 3600 / 24 / 365.25)
@@ -28,6 +29,22 @@ function evaluate_powerlaw_red_noise_gp(log10_A, γ, αs, βs, f1, Δt, ln_js)
     return σ1 * result
 end
 
+function evaluate_powerlaw_red_noise_weights_inv(log10_A, γ, f1, ln_js)
+    A = exp10(log10_A)
+
+    P1 = powerlaw(A, γ, f1, f1)
+    @assert udim(P1) == 2
+    P1 = value(P1)
+
+    npar = length(ln_js)
+    weights_inv = Vector{Float64}(undef, 2 * npar)
+    @inbounds for p = 1:npar
+        weights_inv[p] = weights_inv[p+npar] = 1 / (P1 * exp(-γ * ln_js[p]))
+    end
+
+    return weights_inv
+end
+
 """
     PowerlawRedNoiseGP
 
@@ -37,8 +54,11 @@ power spectral density is assumed to be a power law.
 struct PowerlawRedNoiseGP{N} <: RedNoiseBase
     ln_js::NTuple{N,Float32}
 
-    PowerlawRedNoiseGP(N::Int) = new{N}(Tuple(map(Float32 ∘ log, 1:N)))
+    PowerlawRedNoiseGP(N::Int) = new{N}(Tuple(map(log, 1:N)))
 end
+
+is_gp_noise(::PowerlawRedNoiseGP) = true
+get_gp_npars(arn::PowerlawRedNoiseGP) = 2 * length(arn.ln_js)
 
 delay(arn::PowerlawRedNoiseGP, toa::TOA, toacorr::TOACorrection, params::NamedTuple) =
     evaluate_powerlaw_red_noise_gp(
@@ -51,6 +71,17 @@ delay(arn::PowerlawRedNoiseGP, toa::TOA, toacorr::TOACorrection, params::NamedTu
         arn.ln_js,
     )
 
+calc_noise_weights_inv(arn::PowerlawRedNoiseGP, params::NamedTuple) =
+    evaluate_powerlaw_red_noise_weights_inv(
+        params.TNREDAMP,
+        params.TNREDGAM,
+        params.PLREDFREQ,
+        arn.ln_js,
+    )
+
+show(io::IO, arn::PowerlawRedNoiseGP) =
+    print(io, "PowerlawRedNoiseGP($(length(arn.ln_js)) harmonics)")
+
 """
     PowerlawDispersionNoiseGP
 
@@ -62,6 +93,9 @@ struct PowerlawDispersionNoiseGP{N} <: DispersionNoiseBase
 
     PowerlawDispersionNoiseGP(N::Int) = new{N}(Tuple(map(log, 1:N)))
 end
+
+is_gp_noise(::PowerlawDispersionNoiseGP) = true
+get_gp_npars(dmn::PowerlawDispersionNoiseGP) = 2 * length(dmn.ln_js)
 
 function dispersion_slope(
     dmn::PowerlawDispersionNoiseGP,
@@ -81,6 +115,17 @@ function dispersion_slope(
     )
 end
 
+calc_noise_weights_inv(dmn::PowerlawDispersionNoiseGP, params::NamedTuple) =
+    evaluate_powerlaw_red_noise_weights_inv(
+        params.TNDMAMP,
+        params.TNDMGAM,
+        params.PLDMFREQ,
+        dmn.ln_js,
+    )
+
+show(io::IO, dmn::PowerlawDispersionNoiseGP) =
+    print(io, "PowerlawDispersionNoiseGP($(length(dmn.ln_js)) harmonics)")
+
 """
     PowerlawDispersionNoiseGP
 
@@ -92,6 +137,9 @@ struct PowerlawChromaticNoiseGP{N} <: ChromaticNoiseBase
 
     PowerlawChromaticNoiseGP(N::Int) = new{N}(Tuple(map(log, 1:N)))
 end
+
+is_gp_noise(::PowerlawChromaticNoiseGP) = true
+get_gp_npars(crn::PowerlawChromaticNoiseGP) = 2 * length(crn.ln_js)
 
 function chromatic_slope(
     crn::PowerlawChromaticNoiseGP,
@@ -110,3 +158,14 @@ function chromatic_slope(
         crn.ln_js,
     )
 end
+
+calc_noise_weights_inv(crn::PowerlawChromaticNoiseGP, params::NamedTuple) =
+    evaluate_powerlaw_red_noise_weights_inv(
+        params.TNCHROMAMP,
+        params.TNCHROMGAM,
+        params.PLCHROMFREQ,
+        crn.ln_js,
+    )
+
+show(io::IO, crn::PowerlawChromaticNoiseGP) =
+    print(io, "PowerlawChromaticNoiseGP($(length(crn.ln_js)) harmonics)")